feat: 폴더 공유 모달에 폴더명 표시
